requirements.txt
nltk==3.7
numpy==1.18.5
pandas==1.3.5
Pillow==9.1.0
scikit_learn==1.0.2
streamlit==1.3.1
tensorflow==2.3.4
